requirements.txt
torch
pillow
tensorflow==2.8.0
matplotlib
torchvision
PyYAML
uvicorn
fastapi
opencv-python
numpy
starlette
google
vietocr
protobuf==3.20.0
python-multipart